L1-support-tools/07-ad-provisioning-cli/macos/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ad-provision"
version = "1.0.0"
description = "CLI for Azure AD/EntraID user provisioning"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "L1 Support Team"}
]
dependencies = [
    "click>=8.0",
    "msgraph-sdk>=1.0",
    "azure-identity>=1.14",
    "python-dotenv>=1.0",
]

[project.scripts]
ad-provision = "ad_provision.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["ad_provision*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
